(chore: naming) renamed `utils` to `sys_commands` to better represent the content. Moved some content outside of it.
