+++ mopro-ffi/src/app_config/sys_commands.rs
@@ -1,7 +1,6 @@
-use std::path::{Path, PathBuf};
+use std::io;
+use std::path::Path;
 use std::process::Command;
-use std::{fs, io};
-use uuid::Uuid;
 
 pub fn build_cdylib(build_dir_str: &str) -> io::Result<()> {
     let build_dir = Path::new(build_dir_str);
@@ -22,28 +21,6 @@ pub fn build_cdylib(build_dir_str: &str) -> io::Result<()> {
     }
 
     Ok(())
-}
-
-fn tmp_local(build_path: &Path) -> PathBuf {
-    let tmp_path = build_path.join("tmp");
-    if let Ok(metadata) = fs::metadata(&tmp_path) {
-        if !metadata.is_dir() {
-            panic!("non-directory tmp");
-        }
-    } else {
-        fs::create_dir_all(&tmp_path).expect("Failed to create local tmpdir");
-    }
-    tmp_path
-}
-
-pub fn mktemp_local(build_path: &Path) -> PathBuf {
-    let dir = tmp_local(build_path).join(&Uuid::new_v4().to_string());
-    fs::create_dir(&dir).expect("Failed to create tmpdir");
-    dir
-}
-
-pub fn cleanup_tmp_local(build_path: &Path) {
-    fs::remove_dir_all(tmp_local(build_path)).expect("Failed to remove tmpdir");
 }
 
 pub fn install_ndk() {
